New repo

+++ rfi_quality_check_vg.py
@@ -8,30 +8,44 @@ from functools import reduce
 import numpy as np
 from time import sleep
 import sys
+from collections import Counter
+from itertools import groupby
+from operator import itemgetter
 
-def rfi_check(base_name, mask_file, time, nchans, tsamp, chanfrac):
+def rfi_check(base_name, mask_file, rfitime, nchans, totime, chanfrac,intfrac):
 	#Create rfifind class and set the channels to zap.
 	a = rfifind_bandpass_on.rfifind(mask_file)
 	channels = float(nchans)
-	intervals = float(tsamp) / float(time)
+	intervals = float(totime) / float(rfitime)
 	a.get_bandpass()
 	# Believed to be unnecessary --> a.set_zap_chans(100.0, 5.0, 2.0, False, 0.01, True, [])
 	zapped = a.mask_zap_chans_per_int
+	# Sometime there are repeted chan numbers in a given interval, so to get unique chan number per interval
+	zapped = [np.unique(x) for x in zapped]	
+
 	int_times = a.times
-	os.system("mv pgplot.ps {0}.ps".format(base_name)) #Rename the .ps file that is outputted from the bandpass plotting function.
-
-	total_zaps = 0
+	if os.path.exists("pgplot.ps"):
+		os.system("mv pgplot.ps {0}.ps".format(base_name)) #Rename the .ps file that is outputted from the bandpass plotting function.
 	
-#	for item in zapped:
-#		total_zaps += item.size
-	zapcount = np.zeros(int(channels))
+	kill_chans = []
 	data = list(np.concatenate(zapped,axis=0))
-		
-	zapcount,edges = np.histogram(data,int(channels))
-	print zapcount 
 
-	#kill_chans = list(reduce(np.intersect1d, zapped))
+	#Get accetable bad intervals for each channel
+	nbadtime = int(intfrac*len(int_times))	
+	cdata = Counter(data)
+
+	for i in range(int(nchans)): 
+		if(cdata[i] > nbadtime): kill_chans.append(i)
+				
 	avg_bad_chans = len(kill_chans) / channels #Give the average fraction of channels flagged across all intervals
+	
+	kill_chans_range = []
+
+	for k,g in groupby(enumerate(kill_chans),lambda (i,x): i-x):
+		temp = map(itemgetter(1),g)
+		kill_chans_range.append(str(str(temp[0])+" "+str(temp[-1])))
+
+	#print kill_chans_range
 
 	chan_killfile_name = base_name + "_chan.kill"
 	chan_killfile = open(chan_killfile_name, "w")
@@ -53,20 +67,29 @@ def rfi_check(base_name, mask_file, time, nchans, tsamp, chanfrac):
 	time_killfile_name = base_name + "_time.kill"
 	time_killfile = open(time_killfile_name, "w")
 	bad_ints = 0
+	kill_time_range = []
 	for i in range(len(int_times)):
-		item = int_times[i]
-		interval = zapped[i]
-		if interval.size / channels >= chanfrac:
-			time_killfile.write("{0}    {1}\n".format(str(item), str(item + time)))
+		#item = int_times[i]
+		#interval = zapped[i]
+		if (zapped[i].size/channels) >= chanfrac:
+			#print i,zapped[i].size/channels,chanfrac
+			temp = "{0}    {1}".format(str(int_times[i]), str(int_times[i] + rfitime))
+			time_killfile.write(temp+"\n")
+			kill_time_range.append((float(int_times[i]),float(int_times[i])+rfitime))
 			bad_ints += 1
+
+	#print kill_time_range
+
 	time_killfile.close()
 
 	percentage_bad_ints = float(bad_ints) / intervals
 
+	total_zaps = bad_ints*len(kill_chans)
+
 	#Calculate the percentage of channel-interval spaces that have been zapped.
 	percent_flagged = (float(total_zaps) / (channels * intervals)) * 100
-	print(str(percent_flagged) + "% of the data was flagged for RFI.")
-	return percent_flagged, percentage_bad_ints
+	print "%.0f%% of the data (%.0f%% chans and %.0f%% interval) was flagged for RFI." % (int(percent_flagged),int(avg_bad_chans*100),int(percentage_bad_ints*100))
+	return percent_flagged, percentage_bad_ints,kill_chans,kill_chans_range,kill_time_range
